pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "chess-mcp-server"
version = "0.1.0"
description = "A Model Context Protocol (MCP) server for playing Chess with AI Agents."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
  { name = "Fritzprix", email = "your-email@example.com" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "mcp[cli]",
    "python-chess",
    "fastapi",
    "uvicorn",
    "jinja2",
]

[project.urls]
Homepage = "https://github.com/your-repo/chess-mcp-server"
Issues = "https://github.com/your-repo/chess-mcp-server/issues"

[project.scripts]
chess-mcp-server = "src.mcp_server:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]
